Introduced mtcp_restart.h.
Renamed mtcp/*.ic to mtcp/*.c.
